Updated where the ALT+Enter command is sent to
Previously the ALT+Enter command was being sent to the active tab,
regardless which tab it was.  This was breaking the refresh if the
Siebel tab wasn't active. Updated to get the tab based on the URL query
so it always sends the ALT+Enter command to the Siebel tab.